pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "msar"
version = "0.1.0"
description = "Multi-Session Access/Refresh authentification manager for FastAPI-based applications"
requires-python = ">=3.13"
authors = [
    { name = "overwhelmed44", email = "grg0410@mail.ru" }
]
dependencies = [
    "fastapi >=0.116.1",
    "makefun >=1.16.0",
    "PyJWT >=2.10.1"
]

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]
